In this commit: (#9)
- Adding JsonToTypedObjectRewriter class which extends Rewriter class and borrows some logic from ScalarFieldToObjectFieldRewriter
- Adding JsonToTypedObjectRewriter class as an export
- Adding JsonToTypedObjectRewriter tests

The purpose of JsonToTypedObjectRewriter is to deal with a situation in which the initial GraphQL schema used a GraphQLJSON type (graphql-type-json) for a field, but there is subsequent desire to type the GraphQLJSON blob in order to strengthen the contract. The issue is that once the client expects a field to be GraphQLJSON, you can't type it out at the GraphQL layer without breaking said client. JsonToTypedObjectRewriter aims to resolve that issue by rewriting the query for the service but still returning the same response as initially expected by the client.