feat(rewriter): expose parent object and element path to rewriter (#20)
* feat(rewriter): expose parent object and element path to rewriter

This changes the implementation of `rewriteResultsAtPath` so that now the parent and path of the element is exposed to `Rewriter`.`rewriteResponse` instead of the element itself. This should support implement field name response rewrites.

BREAKING CHANGE: This changes the definition of `Rewriter`.`rewriteResponse`. The `path` references the location of the element while the `response` points to the parent of the element. Combining these will result in the original behavior.

* refactor: rename property to fix shallowing

* refactor: rename variable for clarity